+++ scanpipe/models.py
@@ -37,6 +37,7 @@ from django.apps import apps
 from django.conf import settings
 from django.core import checks
 from django.core.exceptions import ObjectDoesNotExist
+from django.core.exceptions import ValidationError
 from django.core.serializers.json import DjangoJSONEncoder
 from django.core.validators import EMPTY_VALUES
 from django.db import models
@@ -78,7 +79,6 @@ from rq.job import Job
 from rq.job import JobStatus
 
 from scancodeio import __version__ as scancodeio_version
-from scanpipe import spdx
 from scanpipe import tasks
 
 logger = logging.getLogger(__name__)
@@ -776,8 +776,7 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, models.Model):
 
     def add_pipeline(self, pipeline_name, execute_now=False):
         """
-        Create a new Run instance with the provided `pipeline` on the current
-        project.
+        Create a new Run instance with the provided `pipeline` on the current project.
 
         If `execute_now` is True, the pipeline task is created.
         on_commit() is used to postpone the task creation after the transaction is
@@ -786,10 +785,13 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, models.Model):
         immediately.
         """
         pipeline_class = scanpipe_app.pipelines.get(pipeline_name)
+        if not pipeline_class:
+            raise ValueError(f"Unknown pipeline: {pipeline_name}")
+
         run = Run.objects.create(
             project=self,
             pipeline_name=pipeline_name,
-            description=pipeline_class.get_doc(),
+            description=pipeline_class.get_summary(),
         )
         if execute_now:
             transaction.on_commit(run.execute_task_async)
@@ -986,6 +988,11 @@ class ProjectRelatedQuerySet(
 ):
     def project(self, project):
         return self.filter(project=project)
+
+    def get_or_none(self, *args, **kwargs):
+        """Get the object from provided lookups or get None"""
+        with suppress(self.model.DoesNotExist, ValidationError):
+            return self.get(*args, **kwargs)
 
 
 class ProjectRelatedModel(models.Model):
@@ -1871,6 +1878,8 @@ class CodebaseResource(
 
     def as_spdx(self):
         """Return this CodebaseResource as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         spdx_license_keys = [license["spdx_license_key"] for license in self.licenses]
         copyrights = [copyright["copyright"] for copyright in self.copyrights]
         holders = [holder["holder"] for holder in self.holders]
@@ -2198,6 +2207,8 @@ class DiscoveredPackage(
 
     def as_spdx(self):
         """Return this DiscoveredPackage as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         checksums = [
             spdx.Checksum(algorithm=algorithm, value=checksum_value)
             for algorithm in ["sha1", "md5"]
@@ -2494,6 +2505,8 @@ class DiscoveredDependency(
 
     def as_spdx(self):
         """Return this Package as an SPDX Package entry."""
+        from scanpipe.pipes import spdx
+
         external_refs = []
 
         if package_url := self.package_url: