+++ scanpipe/tests/test_models.py
@@ -464,6 +464,10 @@ class ScanPipeModelsTest(TestCase):
         self.assertEqual(2, project.projecterrors_count)
         self.assertEqual(2, project.projecterrors.count())
 
+    def test_scanpipe_project_related_queryset_get_or_none(self):
+        self.assertIsNone(CodebaseResource.objects.get_or_none(path="path/"))
+        self.assertIsNone(DiscoveredPackage.objects.get_or_none(name="name"))
+
     def test_scanpipe_run_model_set_scancodeio_version(self):
         run1 = Run.objects.create(project=self.project1)
         self.assertEqual("", run1.scancodeio_version)
@@ -623,28 +627,28 @@ class ScanPipeModelsTest(TestCase):
         )
 
         qs = self.project1.runs.has_start_date()
-        self.assertQuerysetEqual(qs, [running, executed, succeed, failed])
+        self.assertQuerySetEqual(qs, [running, executed, succeed, failed])
 
         qs = self.project1.runs.not_started()
-        self.assertQuerysetEqual(qs, [not_started])
+        self.assertQuerySetEqual(qs, [not_started])
 
         qs = self.project1.runs.queued()
-        self.assertQuerysetEqual(qs, [queued])
+        self.assertQuerySetEqual(qs, [queued])
 
         qs = self.project1.runs.running()
-        self.assertQuerysetEqual(qs, [running])
+        self.assertQuerySetEqual(qs, [running])
 
         qs = self.project1.runs.executed()
-        self.assertQuerysetEqual(qs, [executed, succeed, failed])
+        self.assertQuerySetEqual(qs, [executed, succeed, failed])
 
         qs = self.project1.runs.succeed()
-        self.assertQuerysetEqual(qs, [succeed])
+        self.assertQuerySetEqual(qs, [succeed])
 
         qs = self.project1.runs.failed()
-        self.assertQuerysetEqual(qs, [failed])
+        self.assertQuerySetEqual(qs, [failed])
 
         queued_or_running_qs = self.project1.runs.queued_or_running()
-        self.assertQuerysetEqual(queued_or_running_qs, [running, queued])
+        self.assertQuerySetEqual(queued_or_running_qs, [running, queued])
 
     def test_scanpipe_run_model_status_property(self):
         now = timezone.now()
@@ -1091,19 +1095,19 @@ class ScanPipeModelsTest(TestCase):
         resource_qs = self.project1.codebaseresources
 
         categories = ["Permissive"]
-        self.assertQuerysetEqual([], resource_qs.licenses_categories(categories))
+        self.assertQuerySetEqual([], resource_qs.licenses_categories(categories))
 
         categories = ["Copyleft"]
         expected = [resource1]
-        self.assertQuerysetEqual(expected, resource_qs.licenses_categories(categories))
+        self.assertQuerySetEqual(expected, resource_qs.licenses_categories(categories))
 
         categories = ["Copyleft Limited"]
         expected = [resource2]
-        self.assertQuerysetEqual(expected, resource_qs.licenses_categories(categories))
+        self.assertQuerySetEqual(expected, resource_qs.licenses_categories(categories))
 
         categories = ["Copyleft", "Copyleft Limited"]
         expected = [resource1, resource2]
-        self.assertQuerysetEqual(expected, resource_qs.licenses_categories(categories))
+        self.assertQuerySetEqual(expected, resource_qs.licenses_categories(categories))
 
     def _create_resources_for_queryset_methods(self):
         resource1 = CodebaseResource.objects.create(project=self.project1, path="1")
@@ -1129,30 +1133,30 @@ class ScanPipeModelsTest(TestCase):
         resource1, resource2, resource3 = self._create_resources_for_queryset_methods()
 
         qs = CodebaseResource.objects
-        self.assertQuerysetEqual([resource2], qs.json_field_contains("holders", "H3"))
-        self.assertQuerysetEqual([resource1], qs.json_field_contains("holders", "H1"))
+        self.assertQuerySetEqual([resource2], qs.json_field_contains("holders", "H3"))
+        self.assertQuerySetEqual([resource1], qs.json_field_contains("holders", "H1"))
         expected = [resource1, resource2]
-        self.assertQuerysetEqual(expected, qs.json_field_contains("holders", "H"))
+        self.assertQuerySetEqual(expected, qs.json_field_contains("holders", "H"))
 
     def test_scanpipe_codebase_resource_queryset_json_list_contains(self):
         resource1, resource2, resource3 = self._create_resources_for_queryset_methods()
         qs = CodebaseResource.objects
 
         results = qs.json_list_contains("holders", "holder", ["H3"])
-        self.assertQuerysetEqual([resource2], results)
+        self.assertQuerySetEqual([resource2], results)
 
         results = qs.json_list_contains("holders", "holder", ["H1"])
-        self.assertQuerysetEqual([resource1], results)
+        self.assertQuerySetEqual([resource1], results)
         results = qs.json_list_contains("holders", "holder", ["H2"])
-        self.assertQuerysetEqual([resource1], results)
+        self.assertQuerySetEqual([resource1], results)
         results = qs.json_list_contains("holders", "holder", ["H1", "H2"])
-        self.assertQuerysetEqual([resource1], results)
+        self.assertQuerySetEqual([resource1], results)
 
         results = qs.json_list_contains("holders", "holder", ["H1", "H2", "H3"])
-        self.assertQuerysetEqual([resource1, resource2], results)
+        self.assertQuerySetEqual([resource1, resource2], results)
 
         results = qs.json_list_contains("holders", "holder", ["H"])
-        self.assertQuerysetEqual([], results)
+        self.assertQuerySetEqual([], results)
 
     def test_scanpipe_codebase_resource_queryset_values_from_json_field(self):
         CodebaseResource.objects.all().delete()
@@ -1178,7 +1182,7 @@ class ScanPipeModelsTest(TestCase):
         CodebaseResource.objects.all().delete()
         self._create_resources_for_queryset_methods()
         results = CodebaseResource.objects.most_common_values("mime_type", limit=1)
-        self.assertQuerysetEqual(["application/zip"], results)
+        self.assertQuerySetEqual(["application/zip"], results)
 
     def test_scanpipe_codebase_resource_queryset_less_common_values(self):
         CodebaseResource.objects.all().delete()
@@ -1188,7 +1192,8 @@ class ScanPipeModelsTest(TestCase):
         )
 
         results = CodebaseResource.objects.less_common_values("mime_type", limit=1)
-        self.assertQuerysetEqual(["text/plain", "text/x-script.python"], results)
+        expected = ["text/plain", "text/x-script.python"]
+        self.assertQuerySetEqual(expected, results, ordered=False)
 
     def test_scanpipe_codebase_resource_queryset_less_common(self):
         CodebaseResource.objects.all().delete()
@@ -1206,10 +1211,10 @@ class ScanPipeModelsTest(TestCase):
 
         qs = CodebaseResource.objects
         results = qs.less_common("mime_type", limit=1)
-        self.assertQuerysetEqual([resource3, resource4], results)
+        self.assertQuerySetEqual([resource3, resource4], results)
 
         results = qs.less_common("holders", limit=2)
-        self.assertQuerysetEqual([resource2], results)
+        self.assertQuerySetEqual([resource2], results)
 
     def test_scanpipe_codebase_resource_descendants(self):
         path = "asgiref-3.3.0-py3-none-any.whl-extract/asgiref"
@@ -1522,14 +1527,19 @@ class ScanPipeModelsTransactionTest(TransactionTestCase):
 
         self.assertEqual(0, project1.runs.count())
 
-        pipeline_name = "docker"
-        pipeline_class = scanpipe_app.pipelines.get(pipeline_name)
+        pipeline_name = "not_available"
+        with self.assertRaises(ValueError) as error:
+            project1.add_pipeline(pipeline_name)
+        self.assertEqual("Unknown pipeline: not_available", str(error.exception))
+
+        pipeline_name = "inspect_manifest"
         project1.add_pipeline(pipeline_name)
+        pipeline_class = scanpipe_app.pipelines.get(pipeline_name)
 
         self.assertEqual(1, project1.runs.count())
         run = project1.runs.get()
         self.assertEqual(pipeline_name, run.pipeline_name)
-        self.assertEqual(pipeline_class.get_doc(), run.description)
+        self.assertEqual(pipeline_class.get_summary(), run.description)
         mock_execute_task.assert_not_called()
 
         project1.add_pipeline(pipeline_name, execute_now=True)
@@ -1600,7 +1610,7 @@ class ScanPipeModelsTransactionTest(TransactionTestCase):
         self.assertEqual("DiscoveredPackage", error.model)
         self.assertEqual(long_value, error.details["filename"])
         self.assertEqual(
-            "value too long for type character varying(255)\n", error.message
+            "value too long for type character varying(255)", error.message
         )
 
         codebase_resource = CodebaseResource.objects.create(
@@ -1733,7 +1743,7 @@ class ScanPipeModelsTransactionTest(TransactionTestCase):
         self.assertEqual(package_count, DiscoveredPackage.objects.count())
         error = project1.projecterrors.latest("created_date")
         self.assertEqual("DiscoveredPackage", error.model)
-        expected_message = "value too long for type character varying(100)\n"
+        expected_message = "value too long for type character varying(100)"
         self.assertEqual(expected_message, error.message)
         self.assertEqual(bad_data["version"], error.details["version"])
         self.assertTrue(error.details["codebase_resource_pk"])